Update SPOTgreedy.py
mincostvalues are not integers

+++ python/SPOTgreedy.py
@@ -22,7 +22,7 @@ def SPOT_GreedySubsetSelection(C, targetMarginal, m):
     setValues = cp.zeros((1, m), dtype=int)
     sizeS = 0
     currOptw = []
-    currMinCostValues = cp.ones((1, numX), dtype=int) * 1000000
+    currMinCostValues = cp.ones((1, numX)) * 1000000
     currMinSourceIndex = cp.zeros((1, numX), dtype=int)
     remainingElements = allY
     chosenElements = []
